Prevent duplicate items from being created/requested. Refactor Cart Items behavior
